citra-qt: Move config dialog code to its own directory
