+++ src/citra_qt/configuration/configure_debug.cpp
@@ -2,7 +2,7 @@
 // Licensed under GPLv2 or any later version
 // Refer to the license.txt file included.
 
-#include "citra_qt/configure_debug.h"
+#include "citra_qt/configuration/configure_debug.h"
 #include "core/settings.h"
 #include "ui_configure_debug.h"
