+++ src/citra_qt/configuration/configure_debug.ui
