+++ src/citra_qt/configuration/configure_input.ui
