+++ src/citra_qt/configuration/configure_system.ui
